// tijarah-pos-main/src/types/entity-name.ts
enum EntityNames {
  BusinessDetailsPull = "business-details-pull",
  ProductsPull = "products-pull",
  CategoryPull = "category-pull",
  CollectionPull = "collection-pull",
  CustomerPull = "customer-pull",
  OrdersPull = "orders-pull",
  PrintTemplatePull = "print-template-pull",
  StockHistoryPull = "stock-history-pull",
  BatchPull = "batch-pull",
  QuickItemsPull = "quick-items-pull",
  CustomChargePull = "custom-charge-pull",
  ProductsPush = "products-push",
  CustomerPush = "customer-push",
  BusinessDetailsPush = "business-details-push",
  OrdersPush = "orders-push",
  PrintTemplatePush = "print-template-push",
  CashDrawerTxnsPush = "cash-drawer-txns-push",
  StockHistoryPush = "stock-history-push",
  BatchPush = "batch-push",
  QuickItemsPush = "quick-items-push",
  BillingSettingsPush = "billing-settings-push",
  BillingSettingsPull = "billing-settings-pull",
  AdsManagementPull = "ads-management-pull",
  AdsReportPush = "ads-report-push",
  KitchenManagementPull = "kitchen-management-pull",
  KitchenManagementPush = "kitchen-management-push",
  SectionTablesPull = "section-tables-pull",
  MenuPull = "menu-pull",
  SectionTablesPush = "section-tables-push",
  VoidCompPull = "void-comp-pull",
  BoxCratesPull = "box-crates-pull",
  BoxCratesPush = "box-crates-push",
}

export default EntityNames;